Move azfs to core, and combine into less files (for easy work on Windows build)
