--- tensorflow_io/core/azure/azfs/azfs_ops.cc
@@ -1,28 +1,0 @@
-/* Copyright 2019 The TensorFlow Authors. All Rights Reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-==============================================================================*/
-
-#include "tensorflow/core/platform/env.h"
-
-#include "tensorflow_io/core/azure/azfs/azfs.h"
-
-namespace tensorflow {
-namespace io {
-namespace {
-
-REGISTER_FILE_SYSTEM("az", io::AzBlobFileSystem);
-
-}  // namespace
-}  // namespace io
-}  // namespace tensorflow
