+++ tensorflow_io/core/kernels/azfs_kernels.cc
@@ -12,28 +12,362 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
+#include "tensorflow_io/core/kernels/azfs_kernels.h"
+#include "tensorflow/core/lib/io/path.h"
+#include "tensorflow/core/lib/strings/str_util.h"
+#include "tensorflow/core/platform/env.h"
+
+#include "blob/blob_client.h"
+#include "logging.h"
+#include "storage_account.h"
+#include "storage_credential.h"
+#include "storage_errno.h"
+
 #include <algorithm>
 #include <chrono>
 #include <fstream>
 #include <ostream>
 #include <sstream>
 
-#include "tensorflow/core/lib/io/path.h"
-#include "tensorflow/core/lib/strings/str_util.h"
-#include "tensorflow/core/platform/env.h"
-#include "tensorflow_io/core/azure/azfs/azfs.h"
-#include "tensorflow_io/core/azure/azfs/azfs_client.h"
-#include "tensorflow_io/core/azure/azfs/azfs_random_access_file.h"
-#include "tensorflow_io/core/azure/azfs/azfs_readonly_memory_region.h"
-#include "tensorflow_io/core/azure/azfs/azfs_writable_file.h"
-
-#include "blob/blob_client.h"
-#include "storage_account.h"
-#include "storage_credential.h"
-#include "storage_errno.h"
-
 namespace tensorflow {
 namespace io {
+namespace {
+// TODO: DO NOT use a hardcoded path
+Status GetTmpFilename(std::string *filename) {
+  if (!filename) {
+    return errors::Internal("'filename' cannot be nullptr.");
+  }
+#ifndef _WIN32
+  char buffer[] = "/tmp/az_blob_filesystem_XXXXXX";
+  int fd = mkstemp(buffer);
+  if (fd < 0) {
+    return errors::Internal("Failed to create a temporary file.");
+  }
+#else
+  char buffer[] = "/tmp/az_blob_filesystem_XXXXXX";
+  char *ret = _mktemp(buffer);
+  if (ret == nullptr) {
+    return errors::Internal("Failed to create a temporary file.");
+  }
+#endif
+  *filename = buffer;
+  return Status::OK();
+}
+
+constexpr char kAzBlobEndpoint[] = ".blob.core.windows.net";
+
+/// \brief Splits a Azure path to a account, container and object.
+///
+/// For example,
+/// "az://account-name.blob.core.windows.net/container/path/to/file.txt" gets
+/// split into "account-name", "container" and "path/to/file.txt".
+Status ParseAzBlobPath(StringPiece fname, bool empty_object_ok,
+                       std::string *account, std::string *container,
+                       std::string *object) {
+  if (!account || !object) {
+    return errors::Internal("account and object cannot be null.");
+  }
+  StringPiece scheme, accountp, objectp;
+  io::ParseURI(fname, &scheme, &accountp, &objectp);
+  if (scheme != "az") {
+    return errors::InvalidArgument(
+        "Azure Blob Storage path doesn't start with 'az://': ", fname);
+  }
+
+  // Consume blob.core.windows.net if it exists
+  absl::ConsumeSuffix(&accountp, kAzBlobEndpoint);
+
+  if (accountp.empty() || accountp.compare(".") == 0) {
+    return errors::InvalidArgument(
+        "Azure Blob Storage path doesn't contain a account name: ", fname);
+  }
+
+  *account = std::string(accountp);
+
+  absl::ConsumePrefix(&objectp, "/");
+
+  auto pos = objectp.find('/');
+  if (pos == std::string::npos) {
+    *container = objectp.data();
+    *object = "";
+  } else {
+    *container = std::string(objectp.substr(0, pos));
+    *object = std::string(objectp.substr(pos + 1));
+  }
+
+  return Status::OK();
+}
+
+std::string errno_to_string() {
+  switch (errno) {
+    /* common errors */
+    case invalid_parameters:
+      return "invalid_parameters";
+    /* client level */
+    case client_init_fail:
+      return "client_init_fail";
+    case client_already_init:
+      return "client_already_init";
+    case client_not_init:
+      return "client_not_init";
+    /* container level */
+    case container_already_exists:
+      return "container_already_exists";
+    case container_not_exists:
+      return "container_not_exists";
+    case container_name_invalid:
+      return "container_name_invalid";
+    case container_create_fail:
+      return "container_create_fail";
+    case container_delete_fail:
+      return "container_delete_fail";
+    /* blob level */
+    case blob__already_exists:
+      return "blob__already_exists";
+    case blob_not_exists:
+      return "blob_not_exists";
+    case blob_name_invalid:
+      return "blob_name_invalid";
+    case blob_delete_fail:
+      return "blob_delete_fail";
+    case blob_list_fail:
+      return "blob_list_fail";
+    case blob_copy_fail:
+      return "blob_copy_fail";
+    case blob_no_content_range:
+      return "blob_no_content_range";
+    /* unknown error */
+    case unknown_error:
+    default:
+      return "unknown_error - " + std::to_string(errno);
+  }
+}
+
+std::shared_ptr<azure::storage_lite::storage_credential> get_credential(
+    const std::string &account) {
+  const auto key = std::getenv("TF_AZURE_STORAGE_KEY");
+  if (key != nullptr) {
+    return std::make_shared<azure::storage_lite::shared_key_credential>(account,
+                                                                        key);
+  } else {
+    return std::make_shared<azure::storage_lite::anonymous_credential>();
+  }
+}
+
+azure::storage_lite::blob_client_wrapper CreateAzBlobClientWrapper(
+    const std::string &account) {
+  azure::storage_lite::logger::set_logger(
+      [](azure::storage_lite::log_level level, const std::string &log_msg) {
+        switch (level) {
+          case azure::storage_lite::log_level::info:
+            _TF_LOG_INFO << log_msg;
+            break;
+          case azure::storage_lite::log_level::error:
+          case azure::storage_lite::log_level::critical:
+            _TF_LOG_ERROR << log_msg;
+            break;
+          case azure::storage_lite::log_level::warn:
+            _TF_LOG_WARNING << log_msg;
+            break;
+          case azure::storage_lite::log_level::trace:
+          case azure::storage_lite::log_level::debug:
+          default:
+            break;
+        }
+      });
+
+  const auto use_dev_account = std::getenv("TF_AZURE_USE_DEV_STORAGE");
+  if (use_dev_account != nullptr) {
+    auto storage_account =
+        azure::storage_lite::storage_account::development_storage_account();
+    auto blob_client =
+        std::make_shared<azure::storage_lite::blob_client>(storage_account, 10);
+    azure::storage_lite::blob_client_wrapper blob_client_wrapper(blob_client);
+    return blob_client_wrapper;
+  }
+
+  const auto use_http_env = std::getenv("TF_AZURE_STORAGE_USE_HTTP");
+  const auto use_https = use_http_env == nullptr;
+  const auto blob_endpoint =
+      std::string(std::getenv("TF_AZURE_STORAGE_BLOB_ENDPOINT") ?: "");
+
+  auto credentials = get_credential(account);
+  auto storage_account = std::make_shared<azure::storage_lite::storage_account>(
+      account, credentials, use_https, blob_endpoint);
+  auto blob_client =
+      std::make_shared<azure::storage_lite::blob_client>(storage_account, 10);
+  azure::storage_lite::blob_client_wrapper blob_client_wrapper(blob_client);
+
+  return blob_client_wrapper;
+}
+
+class AzBlobRandomAccessFile : public RandomAccessFile {
+ public:
+  AzBlobRandomAccessFile(const std::string &account,
+                         const std::string &container,
+                         const std::string &object)
+      : account_(account), container_(container), object_(object) {}
+  ~AzBlobRandomAccessFile() {}
+  Status Read(uint64 offset, size_t n, StringPiece *result,
+              char *scratch) const override {
+    // If n == 0, then return Status::OK()
+    // otherwise, if bytes_read < n then return OutofRange
+    if (n == 0) {
+      *result = StringPiece("", 0);
+      return Status::OK();
+    }
+    auto blob_client = CreateAzBlobClientWrapper(account_);
+    auto blob_property = blob_client.get_blob_property(container_, object_);
+    if (errno != 0) {
+      return errors::Internal("Failed to get properties");
+    }
+    int64 file_size = blob_property.size;
+
+    size_t bytes_to_read = n;
+    if (offset >= file_size) {
+      bytes_to_read = 0;
+    } else if (offset + n > file_size) {
+      bytes_to_read = file_size - offset;
+    }
+
+    if (bytes_to_read == 0) {
+      *result = StringPiece("", 0);
+    } else {
+      std::ostringstream oss;
+      // https://stackoverflow.com/a/12481580
+#ifndef __APPLE__
+      oss.rdbuf()->pubsetbuf(scratch, bytes_to_read);
+#endif
+
+      blob_client.download_blob_to_stream(container_, object_, offset,
+                                          bytes_to_read, oss);
+      if (errno != 0) {
+        *result = StringPiece("", 0);
+        return errors::Internal("Failed to get contents of az://", account_,
+                                kAzBlobEndpoint, "/", container_, "/", object_,
+                                " (", errno_to_string(), ")");
+      }
+
+#ifndef __APPLE__
+      *result = StringPiece(scratch, bytes_to_read);
+#else
+      auto blob_string = oss.str();
+      if (scratch != nullptr) {
+        std::copy(blob_string.begin(), blob_string.end(), scratch);
+      }
+      *result = StringPiece(blob_string);
+#endif
+    }
+    if (bytes_to_read < n) {
+      return errors::OutOfRange("EOF reached");
+    }
+
+    return Status::OK();
+  }
+
+ private:
+  std::string account_;
+  std::string container_;
+  std::string object_;
+};
+
+class AzBlobWritableFile : public WritableFile {
+ public:
+  AzBlobWritableFile(const std::string &account, const std::string &container,
+                     const std::string &object)
+      : account_(account),
+        container_(container),
+        object_(object),
+        sync_needed_(true) {
+    if (GetTmpFilename(&tmp_content_filename_).ok()) {
+      outfile_.open(tmp_content_filename_,
+                    std::ofstream::binary | std::ofstream::app);
+    }
+  }
+
+  ~AzBlobWritableFile() { Close().IgnoreError(); }
+
+  Status Append(StringPiece data) override {
+    TF_RETURN_IF_ERROR(CheckWritable());
+    sync_needed_ = true;
+    outfile_ << data;
+    if (!outfile_.good()) {
+      return errors::Internal(
+          "Could not append to the internal temporary file.");
+    }
+    return Status::OK();
+  }
+
+  Status Close() override {
+    if (outfile_.is_open()) {
+      TF_RETURN_IF_ERROR(Sync());
+      outfile_.close();
+      std::remove(tmp_content_filename_.c_str());
+    }
+    return Status::OK();
+  }
+
+  Status Flush() override { return Sync(); }
+  Status Sync() override {
+    TF_RETURN_IF_ERROR(CheckWritable());
+    if (!sync_needed_) {
+      return Status::OK();
+    }
+    const auto status = SyncImpl();
+    if (status.ok()) {
+      sync_needed_ = false;
+    }
+    return status;
+  }
+
+ private:
+  Status SyncImpl() {
+    outfile_.flush();
+    if (!outfile_.good()) {
+      return errors::Internal(
+          "Could not write to the internal temporary file.");
+    }
+
+    auto blob_client = CreateAzBlobClientWrapper(account_);
+    blob_client.upload_file_to_blob(tmp_content_filename_, container_, object_);
+    if (errno != 0) {
+      return errors::Internal("Failed to upload to az://", account_, "/",
+                              container_, "/", object_, " (", errno_to_string(),
+                              ")");
+    }
+
+    return Status::OK();
+  }
+
+  Status CheckWritable() const {
+    if (!outfile_.is_open()) {
+      return errors::FailedPrecondition(
+          "The internal temporary file is not writable.");
+    }
+    return Status::OK();
+  }
+
+  std::string account_;
+  std::string container_;
+  std::string object_;
+  std::string tmp_content_filename_;
+  std::ofstream outfile_;
+  bool sync_needed_;  // whether there is buffered data that needs to be synced
+};
+
+class AzBlobReadOnlyMemoryRegion : public ReadOnlyMemoryRegion {
+ public:
+  AzBlobReadOnlyMemoryRegion(std::unique_ptr<char[]> data, uint64 length)
+      : data_(std::move(data)), length_(length) {}
+  const void *data() override { return reinterpret_cast<void *>(data_.get()); }
+  uint64 length() override { return length_; }
+
+ private:
+  std::unique_ptr<char[]> data_;
+  uint64 length_;
+};
+
+}  // namespace
 
 Status AzBlobFileSystem::NewRandomAccessFile(
     const std::string &filename, std::unique_ptr<RandomAccessFile> *result) {
@@ -473,6 +807,8 @@ Status AzBlobFileSystem::ListResources(
 
   return Status::OK();
 }
-
+namespace {
+REGISTER_FILE_SYSTEM("az", io::AzBlobFileSystem);
+}  // namespace
 }  // namespace io
 }  // namespace tensorflow
